+++ powerbox/__init__.py
@@ -1,5 +1,5 @@
 __version__ = "0.5.0"
 
-from powerbox import PowerBox, LogNormalPowerBox
-from tools import get_power, angular_average
-from dft import fft, ifft, fftfreq
+from .powerbox import PowerBox, LogNormalPowerBox
+from .tools import get_power, angular_average
+from .dft import fft, ifft, fftfreq