+++ tests/test_power.py
@@ -14,7 +14,7 @@ def test_power1d():
         pb = PowerBox(8001, dim=1, pk=lambda k: 1.0*k ** -2., boxlength=1.0, a=0,b=1 )
         p[i], k = get_power(pb.delta_x(), pb.boxlength,a=0,b=1)
 
-    print np.mean(np.array(p),axis=0)/(1.0*k**-2.)
+    print(np.mean(np.array(p),axis=0)/(1.0*k**-2.))
     assert np.allclose(np.mean(np.array(p),axis=0)[2000:],1.0*k[2000:]**-2.,rtol=2)
 
 def test_power1d_n3():
@@ -23,7 +23,7 @@ def test_power1d_n3():
         pb = PowerBox(8001, dim=1, pk=lambda k: 1.0*k ** -3., boxlength=1.0, b=1 )
         p[i], k = get_power(pb.delta_x(), pb.boxlength,b=1)
 
-    print np.mean(np.array(p), axis=0)/(1.0*k ** -2.)
+    print(np.mean(np.array(p), axis=0)/(1.0*k ** -2.))
     assert np.allclose(np.mean(np.array(p), axis=0)[2000:], 1.0*k[2000:] ** -2., rtol=2)
 
 
@@ -33,7 +33,7 @@ def test_power1d_bigL():
         pb = PowerBox(8001, dim=1, pk=lambda k: 1.0*k ** -3., boxlength=10.0, b=1 )
         p[i], k = get_power(pb.delta_x(), pb.boxlength,b=1)
 
-    print np.mean(np.array(p), axis=0)/(1.0*k ** -2.)
+    print(np.mean(np.array(p), axis=0)/(1.0*k ** -2.))
     assert np.allclose(np.mean(np.array(p), axis=0)[2000:], 1.0*k[2000:] ** -2., rtol=2)
 
 
@@ -43,7 +43,7 @@ def test_power1d_ordinary_freq():
         pb = PowerBox(8001, dim=1, pk=lambda k: 1.0*k ** -3., boxlength=1.0 )
         p[i], k = get_power(pb.delta_x(), pb.boxlength)
 
-    print np.mean(np.array(p), axis=0)/(1.0*k ** -2.)
+    print(np.mean(np.array(p), axis=0)/(1.0*k ** -2.))
     assert np.allclose(np.mean(np.array(p), axis=0)[2000:], 1.0*k[2000:] ** -2., rtol=2)
 
 
@@ -54,7 +54,7 @@ def test_power1d_halfN():
         pb = PowerBox(4001, dim=1, pk=lambda k: 1.0*k ** -3., boxlength=1.0, b=1 )
         p[i], k = get_power(pb.delta_x(), pb.boxlength,b=1)
 
-    print np.mean(np.array(p), axis=0)/(1.0*k ** -2.)
+    print(np.mean(np.array(p), axis=0)/(1.0*k ** -2.))
     assert np.allclose(np.mean(np.array(p), axis=0)[1000:], 1.0*k[1000:] ** -2., rtol=2)
 
 
@@ -64,7 +64,7 @@ def test_power2d():
         pb = PowerBox(200, dim=2, pk=lambda k: 1.0*k ** -2., boxlength=1.0, b=1)
         p[i], k = get_power(pb.delta_x(), pb.boxlength,b=1)
 
-    print np.mean(np.array(p),axis=0)/(1.0*k**-2.)
+    print(np.mean(np.array(p),axis=0)/(1.0*k**-2.))
     assert np.allclose(np.mean(np.array(p),axis=0)[100:],1.0*k[100:]**-2.,rtol=2)
 
 
@@ -72,5 +72,5 @@ def test_power3d():
     pb = PowerBox(50, dim=2, pk=lambda k: 1.0*k ** -2., boxlength=1.0, b=1)
     p, k = get_power(pb.delta_x(), pb.boxlength,b=1)
 
-    print p/(1.0*k**-2.)
+    print(p/(1.0*k**-2.))
     assert np.allclose(p,1.0*k**-2.,rtol=2)