+++ tests/test_stats.py
@@ -16,5 +16,5 @@ def test_resolution():
     for i in range(6):
         pb = PowerBox(64*2**i, dim=2, pk=lambda k: 1.0*k ** -2., boxlength=1.0, angular_freq=True)
         var[i] = np.var(gaussian_filter(pb.delta_x(),sigma=2**i,mode='wrap'))
-    print var/var[0]
+    print(var/var[0])
     assert np.allclose(var/var[0],1,atol=1e-2)