+++ tests/test_tools.py
@@ -16,7 +16,7 @@ def test_angular_avg_nd_3():
     P = np.repeat(P,100).reshape(400,400,100)
     freq = [x,x,np.linspace(-2,2,100)]
     p_k, k_av_bins,other_k = angular_average_nd(P,freq,bins=50,n=2)
-    print p_k[6:,0], k_av_bins[6:]**-2.
+    print(p_k[6:,0], k_av_bins[6:]**-2.)
     assert np.max(np.abs((p_k[6:,0] - k_av_bins[6:]**-2.)/k_av_bins[6:]**-2.)) <  0.05
 
 
@@ -32,5 +32,5 @@ def test_angular_avg_nd_4_2():
     freq = [x,x,np.linspace(-2,2,10), np.linspace(-2,2,10)]
     p_k, k_av_bins,other_k = angular_average_nd(P,freq,bins=50,n=2)
 
-    print np.abs((p_k[7:,0,0] - k_av_bins[7:]**-2.)/k_av_bins[7:]**-2.)
+    print(np.abs((p_k[7:,0,0] - k_av_bins[7:]**-2.)/k_av_bins[7:]**-2.))
     assert np.max(np.abs((p_k[6:,0,0] - k_av_bins[6:]**-2.)/k_av_bins[6:]**-2.)) <  0.06